todo test

+++ app/Services/Impl/TodolistServiceImpl.php
@@ -26,6 +26,8 @@ class TodolistServiceImpl implements TodolistService
     public function removeTodo(string $todoId)
     {
         $todo = Todo::query()->find($todoId);
-        $todo->delete();
+        if($todo != null){
+            $todo->delete();
+        }
     }
 }
